Add files via upload
In this video, we will make an Age Calculator app with the help of HTML, CSS, and JavaScript. This age calculator app has one data input box, where it displays the calendar to pick the date of birth. after selecting the date of birth click on the calculate button then it will display the exact age till today.